+++ cmake/TargetMacros.cmake
@@ -118,4 +118,47 @@ function(define_scratch_program program_name entry_point_source_name)
 endfunction(define_scratch_program)
 
 
+function(define_cli_program program_name entry_point_source_name tool_name)
+
+	add_executable(${program_name}
+		${entry_point_source_name}
+	)
+
+	# Keep CMake target names distinct from the static library target `libver`.
+	set_target_properties(${program_name} PROPERTIES
+		OUTPUT_NAME "${tool_name}"
+	)
+
+	target_compile_definitions(${program_name}
+		PRIVATE
+			LIBVER_TOOL_NAME=\"${tool_name}\"
+	)
+
+	target_link_libraries(${program_name}
+		PRIVATE
+			libver
+			sistools-common-c::core
+			CLASP::core
+			cstring::core
+			collect-c::core
+			Diagnosticism::core
+			$<$<BOOL:${STLSoft_FOUND}>:STLSoft::STLSoft>
+	)
+
+	if(WIN32)
+
+		target_link_libraries(${program_name}
+			PRIVATE
+				wininet
+		)
+	endif(WIN32)
+
+	define_target_compile_options(${program_name})
+
+	install(TARGETS ${program_name}
+		RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
+	)
+endfunction(define_cli_program)
+
+
 # ############################## end of file ############################# #